Extracted currentUser contexts
