+++ packages/ui/components/providers/CurrentUserProvider.tsx
@@ -1,14 +1,6 @@
-import { createContext } from "react";
-import { KeyedMutator } from "swr";
-import { User } from "lib/types";
 import { useCurrentUser } from "../../hooks/Auth/useCurrentUser";
+import CurrentUserContext from "../../contexts/CurrentUserContext";
 
-export type CurrentUserContextType = {
-  currentUser?: User | undefined;
-  mutate?: KeyedMutator<User | undefined>;
-  error?: Error | undefined;
-};
-export const CurrentUserContext = createContext<CurrentUserContextType>({});
 export const CurrentUserProvider = (props: any) => {
   const { data, mutate, error } = useCurrentUser();
